refactor : 설정 정리, schema 정리
