+++ src/main/java/com/mopl/api/MoplApplication.java
@@ -2,11 +2,7 @@ package com.mopl.api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.scheduling.annotation.EnableScheduling;
 
-@EnableJpaAuditing
-@EnableScheduling
 @SpringBootApplication
 public class MoplApplication {
