+++ src/main/java/com/mopl/api/global/config/JpaConfig.java
@@ -1,0 +1,10 @@
+package com.mopl.api.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@EnableJpaAuditing
+@Configuration
+public class JpaConfig {
+
+}
