+++ src/main/java/com/mopl/api/global/config/QuerydslConfig.java
@@ -2,13 +2,18 @@ package com.mopl.api.global.config;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 public class QuerydslConfig {
+
+    @PersistenceContext
+    private EntityManager em;
+
     @Bean
-    public JPAQueryFactory jpaQueryFactory(EntityManager em) {
+    public JPAQueryFactory jpaQueryFactory() {
         return new JPAQueryFactory(em);
     }
 }
