+++ src/main/java/com/mopl/api/global/config/SchedulingConfig.java
@@ -1,0 +1,10 @@
+package com.mopl.api.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@EnableScheduling
+@Configuration
+public class SchedulingConfig {
+
+}
