+++ src/main/resources/dummy-data.sql
@@ -25,7 +25,7 @@ VALUES (UUID_TO_BIN(UUID()),
         'Inception',
         'A thief who steals corporate secrets through dream-sharing technology is given the inverse task of planting an idea into the mind of a C.E.O.',
         'https://cdn.test/inception.jpg',
-        JSON_ARRAY('sci-fi', 'thriller', 'action'),
+        'sci-fi|thriller|action',
         4.5,
         3,
         150,
@@ -36,7 +36,7 @@ VALUES (UUID_TO_BIN(UUID()),
         'Forrest Gump',
         'The presidencies of Kennedy and Johnson, the Vietnam War, and other historical events unfold from the perspective of an Alabama man.',
         'https://cdn.test/forrest-gump.jpg',
-        JSON_ARRAY('drama', 'romance'),
+        'drama|romance',
         4.8,
         2,
         200,
@@ -47,7 +47,7 @@ VALUES (UUID_TO_BIN(UUID()),
         'Breaking Bad',
         'A high school chemistry teacher turned methamphetamine producer partners with a former student.',
         'https://cdn.test/bb.jpg',
-        JSON_ARRAY('crime', 'drama', 'thriller'),
+        'crime|drama|thriller',
         4.9,
         4,
         320,
@@ -58,7 +58,7 @@ VALUES (UUID_TO_BIN(UUID()),
         'Game of Thrones',
         'Nine noble families fight for control over the lands of Westeros, while an ancient enemy returns.',
         'https://cdn.test/got.jpg',
-        JSON_ARRAY('fantasy', 'drama', 'action'),
+        'fantasy|drama|action',
         4.3,
         5,
         450,
@@ -69,7 +69,7 @@ VALUES (UUID_TO_BIN(UUID()),
         'UEFA Champions League Final 2023',
         'Manchester City vs Inter Milan - The most prestigious club competition in European football.',
         'https://cdn.test/ucl-final.jpg',
-        JSON_ARRAY('football', 'soccer', 'live'),
+        'football|soccer|live',
         4.7,
         2,
         180,
@@ -80,7 +80,7 @@ VALUES (UUID_TO_BIN(UUID()),
         'NBA Finals 2023',
         'Denver Nuggets vs Miami Heat - The championship series of the NBA.',
         'https://cdn.test/nba-finals.jpg',
-        JSON_ARRAY('basketball', 'live', 'sports'),
+        'football|soccer|live',
         4.6,
         1,
         120,
@@ -91,7 +91,7 @@ VALUES (UUID_TO_BIN(UUID()),
         'The Shawshank Redemption',
         'Two imprisoned men bond over a number of years, finding solace and eventual redemption.',
         'https://cdn.test/shawshank.jpg',
-        JSON_ARRAY('drama'),
+        'football|soccer|live',
         4.9,
         6,
         280,
@@ -102,7 +102,7 @@ VALUES (UUID_TO_BIN(UUID()),
         'The Mandalorian',
         'The travels of a lone bounty hunter in the outer reaches of the galaxy.',
         'https://cdn.test/mandalorian.jpg',
-        JSON_ARRAY('sci-fi', 'action', 'adventure'),
+        'football|soccer|live',
         4.4,
         3,
         220,
@@ -728,13 +728,3 @@ FROM (SELECT id, ROW_NUMBER() OVER () as rn FROM conversations) c
          JOIN users s ON s.email = 'charlie@test.com'
          JOIN users r ON r.email = 'bob@test.com'
 WHERE c.rn = 3;
-
--- JWT SESSIONS
-INSERT INTO jwt_sessions (id, user_id, refresh_token, expired_at)
-SELECT UUID_TO_BIN(UUID()),
-       u.id,
-       CONCAT('refresh_token_', u.email, '_', UUID()),
-       DATE_ADD(NOW(), INTERVAL 30 DAY)
-FROM users u
-WHERE u.email IN
-      ('alice@test.com', 'bob@test.com', 'charlie@test.com', 'diana@test.com', 'admin@test.com');
